+++ file_format/yarnRepr.h
@@ -12,11 +12,11 @@ namespace file_format {
 // An internal representation of a yarn in the simulator
 struct Yarn {
   // A list of 3D control points (row based)
-  Eigen::MatrixXf points;
+  Eigen::MatrixXd points;
   // RGB color (in range 0-255)
-  glm::ivec3 color = glm::vec3(128, 128, 128);
+  glm::ivec3 color = glm::ivec3(128, 128, 128);
   // Yarn radius
-  float radius;
+  double radius;
 };
 
 // An internal representation of yarns in the simulator
