+++ simulator/Simulator.Dglobal.cpp
@@ -13,15 +13,15 @@ void Simulator::calculateGlobalDamping(int i) {
 
   DECLARE_POINTS2(pD, dQ, index);
 
-  float coefficient = params.kGlobal;
+  double coefficient = params.kGlobal;
 
-  using Vec12 = Eigen::Matrix<float, 12, 1>;
+  using Vec12 = Eigen::Matrix<double, 12, 1>;
 
-  Vec12 res = integrate<Vec12>([&pD](float s)->Vec12 {
+  Vec12 res = integrate<Vec12>([&pD](double s)->Vec12 {
     Vec12 ans;
     DECLARE_BASIS2(b, s);
 
-    Eigen::Vector3f P = POINT_FROM_BASIS(pD, b);
+    Eigen::Vector3d P = POINT_FROM_BASIS(pD, b);
 
     for (int kk = 0; kk < 4; kk++) {
       ans.block<3, 1>(kk * 3ll, 0) = (2.f * b[kk]) * P;
